autostream_agent/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

langgraph>=0.1.0

faiss-cpu>=1.7.4

openai>=1.0.0
tiktoken>=0.5.0

python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
